refactor: plugin cleanup
