+++ packages/apps/plugins/plugin-client/src/index.ts
@@ -3,3 +3,4 @@
 //
 
 export * from './ClientPlugin';
+export * from './types';
