+++ packages/apps/plugins/plugin-client/src/types.ts
@@ -1,0 +1,10 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+import { Client, ShellController } from '@dxos/react-client';
+
+export type ClientPluginProvides = {
+  client: Client;
+  setLayout: ShellController['setLayout'];
+};
