+++ packages/apps/plugins/plugin-graph/src/GraphPlugin.tsx
@@ -2,77 +2,13 @@
 // Copyright 2023 DXOS.org
 //
 
-import type { IconProps } from '@phosphor-icons/react';
-import React, { UIEvent, FC, createContext, useContext } from 'react';
+import React from 'react';
 
-import type { TFunction } from '@dxos/aurora';
-import { Plugin, PluginDefinition } from '@dxos/react-surface';
+import { PluginDefinition } from '@dxos/react-surface';
 
-export type MaybePromise<T> = T | Promise<T>;
-
-export type GraphNode<TDatum = any> = {
-  id: string;
-  label: string | [string, { ns: string; count?: number }];
-  description?: string;
-  icon?: FC;
-  data?: TDatum; // nit about naming this
-  parent?: GraphNode;
-  children?: GraphNode[];
-  actions?: GraphNodeAction[];
-  attributes?: { [key: string]: any };
-};
-
-export type GraphNodeAction = {
-  id: string;
-  testId?: string;
-  // todo(thure): `Parameters<TFunction>` causes typechecking issues because `TFunction` has so many signatures
-  label: string | [string, { ns: string; count?: number }];
-  icon?: FC<IconProps>;
-  invoke: (t: TFunction, event: UIEvent) => MaybePromise<void>;
-};
-
-export type GraphProvides = {
-  graph: {
-    nodes?: (plugins: Plugin[]) => GraphNode[];
-    actions?: (plugins: Plugin[]) => GraphNodeAction[];
-  };
-};
-
-type GraphPlugin = Plugin<GraphProvides>;
-
-export const findGraphNode = (nodes: GraphNode[], [id, ...path]: string[]): GraphNode | undefined => {
-  const node = nodes.find((n) => n.id === id);
-  if (!node) {
-    return undefined;
-  }
-
-  if (path.length === 0 || !node.children || node.children.length === 0) {
-    return node;
-  }
-
-  return findGraphNode(node.children, path);
-};
-
-export const isGraphNode = (datum: unknown): datum is GraphNode =>
-  datum && typeof datum === 'object' ? 'id' in datum && 'label' in datum : false;
-
-export const graphPlugins = (plugins: Plugin[]): GraphPlugin[] => {
-  return (plugins as GraphPlugin[]).filter((p) => typeof p.provides?.graph?.nodes === 'function');
-};
-
-// TODO(wittjosiah): State can be a GraphNode.
-export type GraphContextValue = {
-  roots: { [key: string]: GraphNode[] };
-  actions: { [key: string]: GraphNodeAction[] };
-};
-
-const GraphContext = createContext<GraphContextValue>({ roots: {}, actions: {} });
-
-export const useGraphContext = () => useContext(GraphContext);
-
-export type GraphPluginProvides = {
-  graph: GraphContextValue;
-};
+import { GraphContext } from './GraphContext';
+import { GraphContextValue, GraphPluginProvides } from './types';
+import { graphPlugins } from './util';
 
 export const GraphPlugin = (
   graph: GraphContextValue = {
@@ -82,7 +18,7 @@ export const GraphPlugin = (
 ): PluginDefinition<GraphPluginProvides> => {
   return {
     meta: {
-      id: 'dxos:GraphPlugin',
+      id: 'dxos:graph',
     },
     ready: async (plugins) => {
       for (const plugin of graphPlugins(plugins)) {
