+++ packages/apps/plugins/plugin-graph/src/index.ts
@@ -2,4 +2,7 @@
 // Copyright 2023 DXOS.org
 //
 
+export * from './GraphContext';
 export * from './GraphPlugin';
+export * from './types';
+export * from './util';
