+++ packages/apps/plugins/plugin-markdown/src/index.ts
@@ -3,4 +3,5 @@
 //
 
 export * from './MarkdownPlugin';
-export * from './props';
+export * from './types';
+export * from './util';
