+++ packages/apps/plugins/plugin-space/src/index.ts
@@ -2,7 +2,6 @@
 // Copyright 2023 DXOS.org
 //
 
-// TODO(wittjosiah): This shouldn't be exported from here.
-export * from './getSpaceDisplayName';
-
 export * from './SpacePlugin';
+export * from './types';
+export * from './util';
