+++ packages/apps/plugins/plugin-theme/src/types.ts
@@ -1,0 +1,9 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+import type { Resource } from 'i18next';
+
+export type TranslationsProvides = {
+  translations: Resource[];
+};
