+++ packages/apps/plugins/plugin-theme/src/util.ts
@@ -1,0 +1,12 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+import { Plugin } from '@dxos/react-surface';
+
+import { TranslationsProvides } from './types';
+
+type TranslationsPlugin = Plugin<TranslationsProvides>;
+export const translationsPlugins = (plugins: Plugin[]): TranslationsPlugin[] => {
+  return (plugins as TranslationsPlugin[]).filter((p) => Array.isArray(p.provides?.translations));
+};
