+++ packages/apps/plugins/plugin-treeview/src/types.ts
@@ -1,0 +1,12 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+// TODO(wittjosiah): Derive graph nodes from selected.
+export type TreeViewContextValue = {
+  selected: string[];
+};
+
+export type TreeViewProvides = {
+  treeView: TreeViewContextValue;
+};
